llm_feature_engineering/requirements.txt
# Core dependencies
hydra-core>=1.3.0
omegaconf>=2.3.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0

# LLM APIs
anthropic>=0.7.0
openai>=1.0.0
python-dotenv>=0.19.0

# AutoGluon (optional)
autogluon.tabular>=0.8.0

# TALENT benchmark (optional)
# Note: Install manually if needed

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950